Switched to Maven and Vaadin 8.
