+++ .gitignore
@@ -4,3 +4,5 @@
 build/
 widgetsets
 gwt-unitCache
+*.iml
+.idea
