+++ assembly/assembly.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<assembly
+	xmlns="http://maven.apache.org/ASSEMBLY/2.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/ASSEMBLY/2.0.0 http://maven.apache.org/xsd/assembly-2.0.0.xsd">
+	<id>addon</id>
+
+	<!-- Note: the base directory for this assembly descriptor is one of the
+		license specific subprojects -->
+
+	<formats>
+		<format>zip</format>
+	</formats>
+
+	<!-- Do not use because we must put META-INF/MANIFEST.MF there. -->
+	<includeBaseDirectory>false</includeBaseDirectory>
+
+	<fileSets>
+		<fileSet>
+			<directory>..</directory>
+			<includes>
+				<include>LICENSE.txt</include>
+				<include>README.md</include>
+			</includes>
+		</fileSet>
+		<fileSet>
+			<directory>target</directory>
+			<outputDirectory/>
+			<includes>
+				<include>*.jar</include>
+                <include>*.pdf</include>
+			</includes>
+		</fileSet>
+	</fileSets>
+
+	<files>
+		<!-- This is vaadin.com/directory related manifest needed in the zip package -->
+		<file>
+			<source>assembly/MANIFEST.MF</source>
+			<outputDirectory>META-INF</outputDirectory>
+			<filtered>true</filtered>
+		</file>
+	</files>
+</assembly>
