+++ src/main/java/org/vaadin/teemu/ratingstars/gwt/RatingStarsWidgetset.gwt.xml
