+++ src/main/java/org/vaadin/teemu/ratingstars/gwt/client/RatingStarsServerRpc.java
