+++ src/main/java/org/vaadin/teemu/ratingstars/gwt/public/ratingstars/images/arrow_head_9x5.png
