+++ src/main/java/org/vaadin/teemu/ratingstars/gwt/public/ratingstars/images/star_12x11.png
