+++ src/main/java/org/vaadin/teemu/ratingstars/gwt/public/ratingstars/images/star_19x18.png
