+++ src/main/java/org/vaadin/teemu/ratingstars/gwt/public/ratingstars/images/star_bg_34x33.gif
