+++ src/main/java/org/vaadin/teemu/ratingstars/gwt/public/ratingstars/styles.css
